+++ ES1Benchmark/ES1Benchmark.h
@@ -49,7 +49,7 @@ namespace Plugin {
 
         // IES1Benchmark
         uint32_t EchoString(const string& value, string& echo) override;
-        uint32_t EchoArray(const std::vector<uint32_t>& values, std::vector<uint32_t>& echo) override;
+        uint32_t EchoArray(const std::vector<uint8_t>& values, std::vector<uint8_t>& echo) override;
         uint32_t EchoUint32(const uint32_t value, uint32_t& echo) override;
         uint32_t EchoUint64(const uint64_t value, uint64_t& echo) override;
         uint32_t EchoBool(const bool value, bool& echo) override;
